refactor: rename button and input components
